requirements/pangeo.txt
# NOTE: These requirements are used for pangeo.

# core packages
cmip6_preprocessing

#testing
pytest
coveralls
recursive_diff     # test netcdfs
typeguard

# linear algebra and general data analysis
numpy               # arrays, linear algebra
scipy               # linear algebra and numerical mathematics
numba               # speeding up array operations
pandas              # tabular data analysis
scipy               # scipy
uncertainties       # linear error propagation

# interactive computing
tqdm                # loading bar

# useful xarray extensions
eccodes
cftime
dask[complete]      # allows more lazy operations for xarray.

# bottleneck          # useful for operations on arrays.
nc-time-axis        # Hopefully this allows time to be decoded
netCDF4             # makes sure that the default driver is netCDF4.

# configuration inputs.
hydra-core

# Wandb
wandb              # automatic recording of runs

# downloads
requests
zipfile38
aiohttp

# intake not working
intake-esm
intake
intake-xarray
gcsfs

# plotting dependencies of package
cmocean
seaborn
